requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-google-genai>=2.0.0
openai>=1.0.0
google-generativeai>=0.8.0
pandas>=2.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
pydantic>=2.0.0
tqdm>=4.60.0
